Adding StaplerGroovyStubsTask - provides 'stapler' task.
Stapler task does generation of groovy stubs and adds the output
directory for the generated stubs to the compileJava task's source
paths, so that Stapler annotations can be processed before groovy compilation.

Note that this does create interim classfiles in the build output
directory based on the stubs, which we may want to clean
up. Alternatively, we could spawn off a new execution of compileJava
for just annotation processing.